epcdata/requirements.txt
asgiref==3.8.1
autopep8==1.5.7
babel==2.17.0
beautifulsoup4==4.13.4
dj-database-url==0.5.0
Django==4.2.23
django-cors-headers==4.7.0
django-countries==7.6.1
django-extensions==4.1
django-extra-views==0.14.0
django-haystack==3.3.0
django-oscar==3.2.4
django-phonenumber-field==6.4.0
django-tables2==2.3.4
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
factory-boy==3.2.1
Faker==37.4.2
gunicorn==20.1.0
openpyxl==3.1.2
packaging==25.0
phonenumbers==9.0.10
pillow==11.2.0
psycopg2-binary==2.9.10
purl==1.6
pycodestyle==2.7.0
PyJWT==2.10.1
python-decouple==3.4
python-dotenv==1.1.1
pytz==2021.1
requests==2.31.0
six==1.17.0
sorl-thumbnail==12.11.0
soupsieve==2.7
sqlparse==0.5.3
toml==0.10.2
typing_extensions==4.14.0
tzdata==2025.1
Unipath==1.1
whitenoise==5.3.0
django-countries
